renderer: Don't recenter viewport when sidebar changes sizes
When the fortress mode sidebar changes sizes, the viewport should remain
fixed at the left side. This behavior of Dwarf Fortress is expected by
various calculations that place the cursor, both in Dwarf Fortress and
in DFHack plugins.

Keep the re-centering behavior for zoom and any other changes, as it is
more friendly than zooming from the top left corner.