CMakeLists.txt: Rely on dfhack/CMake/Common.cmake
This uses CMake's INCLUDE() directive to pull in definitions from the
DFHack project, rather than duplicating them. It's intended to ensure
that the same compiler, essential compiler options, version numbers, and
so forth, are used between DFHack and TWBT. This makes TWBT build more
like a subproject of DFHack.

The LOAD_CACHE() operation is key to getting the configuration as the
user called it when building DFHack. Just doing INCLUDE() will source
the CMake rules, but does not copy any settings set when CMake generated
its build files. But LOAD_CACHE does that, so if needed we can get the
compiler used or other things. To start with, we load the installation
prefix.